requirements.txt
atproto
google-generativeai
python-dotenv
openai
requests
